+++ entity/src/main/resources/db/migration/V1.43__new_column_subject.sql
@@ -1,0 +1,8 @@
+-- ------------------------------------------------
+-- Version: v1.43
+--
+-- Description: Migration that adds new column in Subject table
+-- -------------------------------------------------
+
+ALTER TABLE Subject
+ADD COLUMN created_on datetime(6) DEFAULT NULL;